requirements-dev.txt
# ── Runtime deps (union of all services) ─────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
curl-cffi==0.7.3
selectolax==0.3.21
camoufox[geoip]==0.4.11
google-cloud-firestore==2.16.0
google-cloud-storage==2.16.0
google-cloud-tasks==2.16.3
httpx==0.27.0

# ── Test / dev tools ─────────────────────────────────────────────────────────
pytest==8.2.0
pytest-asyncio==0.23.6
pytest-cov==5.0.0
ruff==0.4.4
coverage==7.5.0
